Preserve path when moving in/out of vm
Vagrant takes an environment variable for the current working directory
so you don't have to change directory to enter the dev vm.

Also... with some ssh commands we can directly go to a particular
directory under ~/govuk.

This means if you are in an app directory on your host machine,
vm will take you to that same directory in the vm, and when
you exit the vm you are back where you were.